feat: Implement CryPlaybackButton for audio playback functionality
